split fixtures file into two files
