+++ client-v2/src/fixtures/frontsConfig.js
@@ -1,34 +1,6 @@
 // @flow
 
-import type { FrontsConfig, FrontsClientConfig } from './types/FrontsConfig';
-
-const liveArticle = {
-  id: 'article/live/0',
-  frontPublicationDate: 1,
-  publishedBy: 'Computers',
-  meta: {}
-};
-
-const draftArticle = {
-  id: 'article/draft/1',
-  frontPublicationDate: 1,
-  publishedBy: 'Computers',
-  meta: {}
-};
-
-const draftArticleInGroup = {
-  id: 'article/draft/2',
-  frontPublicationDate: 1,
-  publishedBy: 'Computers',
-  meta: { group: 1 }
-};
-
-const snapArticle = {
-  id: 'snap/link/3',
-  frontPublicationDate: 1,
-  publishedBy: 'Computers',
-  meta: {}
-};
+import type { FrontsConfig, FrontsClientConfig } from '../types/FrontsConfig';
 
 const frontsConfig: FrontsConfig = {
   fronts: {
@@ -69,11 +41,4 @@ const frontsClientConfig: FrontsClientConfig = {
   }
 };
 
-export {
-  frontsClientConfig,
-  frontsConfig,
-  liveArticle,
-  draftArticle,
-  snapArticle,
-  draftArticleInGroup
-};
+export { frontsClientConfig, frontsConfig };
