+++ client-v2/src/util/__tests__/collectionUtils.spec.js
@@ -2,7 +2,11 @@
 
 import { getCollectionArticleQueryString } from '../collectionUtils';
 import { frontStages } from '../../constants/fronts';
-import { liveArticle, draftArticle, snapArticle } from '../../fixtures';
+import {
+  liveArticle,
+  draftArticle,
+  snapArticle
+} from '../../fixtures/articles';
 
 const collectionWithNoDraftArticles = {
   live: [liveArticle],
